cleanup (issue 36)
