--- hw3app/src/App.css
@@ -1,19 +1,0 @@
-table, th, td{
-  border: 1px solid;
-  font-size: 18px;
-}
-table{
-  width: 60%;
-  margin-top: 48px;
-  background-color: black;
-}
-th, td{
-  color: rebeccapurple;
-  border-color: black;
-  background-color: yellow;
-  padding: 8px;
-}
-
-.rating{
-  text-align: center;
-}
