--- hw3app/src/index.js
@@ -1,21 +1,0 @@
-import ReactDOM from "react-dom/client";
-import React from 'react';
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Layout from "./views/Layout";
-import NoPage from "./views/NoPage";
-import './App.css'
-
-export default function App() {
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Layout />}>
-          <Route path="*" element={<NoPage />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<App />);
