--- hw3app/src/views/Contact.js
@@ -1,7 +1,0 @@
-// Contact.js
-
-const Contact = () => {
-    return <h1>Contact Me</h1>;
-  };
-  
-  export default Contact;
