--- hw3app/src/views/Home.js
@@ -1,7 +1,0 @@
-// Home.js
-
-const Home = () => {
-    return <h1>Home</h1>;
-  };
-  
-  export default Home;
