--- hw3app/src/views/NoPage.js
@@ -1,7 +1,0 @@
-// NoPage.js
-
-const NoPage = () => {
-    return <h1>404</h1>;
-  };
-  
-  export default NoPage;
