+++ src/App.css
@@ -3,7 +3,7 @@ table, th, td{
   font-size: 18px;
 }
 table{
-  width: 60%;
+  width: 100%;
   margin-top: 48px;
   background-color: black;
 }
