+++ src/index.js
@@ -3,7 +3,6 @@ import React from 'react';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "./views/Layout";
 import NoPage from "./views/NoPage";
-import Enumerate from "./App";
 import './App.css'
 
 export default function App() {
@@ -11,7 +10,6 @@ export default function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Layout />}>
-          <Route index element = {<Enumerate />} />
           <Route path="*" element={<NoPage />} />
         </Route>
       </Routes>
