--- src/views/Blogs.js
@@ -1,7 +1,0 @@
-// Blog.js
-
-const Blogs = () => {
-    return <h1>Blog Articles</h1>;
-  };
-  
-  export default Blogs;
